Fix deployment action
The old files from previous deployments weren't getting removed, so we can use the
"rm" option for scp-action to remove the target folder before uploading. The problem is,
that will also remove the .htaccess that's in that folder, so we need to re-upload
the .htaccess we want to the target dir after we remove and redeploy the appp.