Update layout to use statuses and lazy loading
Instead of one mega list that loads every challenge in the DB, we now have three. One for "queued" challenges, one for "active" challenges, and one for "complete" challenges. Each list loads it's own data when it's opened, and only the "active" one starts open, so during the initial page load we only have to wait for the active challenges to load.